+++ src/support/slack/commands/onboard.js
@@ -365,6 +365,7 @@ function OnboardCommand(context) {
           await triggerAuditForSite(
             site,
             auditType,
+            undefined,
             slackContext,
             context,
           );